refactor: split bundle command facade
